Fix sonar problem

+++ dubbo-metrics/dubbo-metrics-api/src/main/java/org/apache/dubbo/metrics/aggregate/TimeWindowCounter.java
@@ -42,8 +42,7 @@ public class TimeWindowCounter {
     }
 
     public long bucketLivedSeconds() {
-        return TimeUnit.MILLISECONDS.toSeconds((long)
-            this.slidingWindow.values().size() * this.slidingWindow.getPaneIntervalInMs());
+        return TimeUnit.MILLISECONDS.toSeconds(this.slidingWindow.values().size() * this.slidingWindow.getPaneIntervalInMs());
     }
 
     public void increment() {
